% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read_ndr.R
\name{current_cols}
\alias{current_cols}
\title{Column types of NDR variables}
\usage{
current_cols()
}
\value{
No return value, called for side effects
}
\description{
Specify the variable types for each of the NDR columns.
}
\keyword{internal}
