% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-summarise_ndr.R
\name{my_summary}
\alias{my_summary}
\title{Utils-summarise-ndr}
\usage{
my_summary(data, l, n)
}
\arguments{
\item{data}{A dataframe argument used by the \code{summarise_ndr()}.}

\item{l}{The level a level at which the summary is to be performed. Options are:
"country", "state" or "facility".}

\item{n}{The name assignment to the summary operator}
}
\description{
Utils-summarise-ndr
}
<<<<<<< HEAD
\keyword{internal}
=======
>>>>>>> 82bc88cea456aae49cd70b9b04969fc72b487d3c
