% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ndr-null.R
\name{ndr-null-op}
\alias{ndr-null-op}
\alias{\%||\%}
\title{Borrowed NULL}
\usage{
x \%||\% y
}
\arguments{
\item{x}{value set to a default value of null}

\item{y}{value to be used with x is null}
}
\description{
Borrowed NULL
}
<<<<<<< HEAD
\keyword{internal}
=======
>>>>>>> 82bc88cea456aae49cd70b9b04969fc72b487d3c
