% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_ndr.R
\name{summarize_ndr}
\alias{summarize_ndr}
\title{Count the number of outcomes based on a specified level}
\usage{
summarize_ndr(data, level, names)
}
\arguments{
\item{data}{A list of dataframes to be summarised}

\item{level}{The level at which the summary should be performed. The options
are "ip" (or "country"), "state" or "facility".}

\item{names}{The names to be passed to the summary columns created in
the output}
}
\value{
summarize_ndr
}
\description{
The \code{summarize_ndr()} function counts the number of occurrence of
specified level for each of the supplied dataframe. It then combines the
given dataframes into a single table. It also adds a "Total" roll that
adds all the rows for each of the numeric columns.
}
\examples{
curr <- tx_curr(ndr_example)
vl_eligible <- tx_vl_eligible(ndr_example)
vl_result <- tx_pvls_den(ndr_example)

summarize_ndr(
  data = list(curr, vl_eligible, vl_result),
  level = "state",
  names = c("tx_curr", "vl_eligible", "tx_pvls_den")
)
}
