% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyndr-package.R
\docType{package}
\name{tidyndr}
\alias{tidyndr}
\title{tidyndr: A package analyzing the Nigeria National Data Repository (NDR)}
\description{
To learn more about tidyndr, refer to \code{browseVignettes(package = "tidyndr")}
}
