% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_curr.R
\name{tx_curr}
\alias{tx_curr}
\title{Subset clients who are on treatment}
\usage{
tx_curr(data, region = unique(data$state), site = unique(data$facility))
}
\arguments{
\item{data}{an ndr dataframe imported using the \code{read_ndr()}.}

\item{region}{a character vector specifying the "State" of interest.
The default utilizes all the states in the dataframe.}

\item{site}{a character vector of at least length 1. Default is to utilize all
the facilities contained in the dataframe.}
}
\value{

}
\description{
Subset clients who are on treatment
}
\examples{
file_path <- "C:/Users/stephenbalogun/Documents/My R/tidyndr/ndr_example.csv"
ndr_example <- read_ndr(file_path)

ndr_example \%>\%
tx_curr()

# generate the TX_CURR for two states (e.g. "State 1" and "State 2" in the ndr_example file)
ndr_example \%>\%
tx_curr(region = c("State 1", "State 2"))

# determine the active clients in two facilities ("Facility 1", and "Facility 2) in "State 1"
ndr_example \%>\%
tx_curr(region = "State 1",
site = c("Facility 1", "Facility 2"))
}
