% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_ml.R
\name{tx_ml}
\alias{tx_ml}
\title{Subset clients who became inactive (IIT) within a given period}
\usage{
tx_ml(
  data,
  from = fy_start,
  to = Sys.Date(),
  region = unique(data$state),
  site = unique(data$facility)
)
}
\arguments{
\item{data}{an ndr dataframe imported using the `read_ndr()}

\item{from}{reference date to start the determination of the "TX_ML"
in ISO8601 format (i.e. "yyyy-mm-dd"). The default is to start at the
beginning of the current Fiscal Year (i.e. 1st of October).}

\item{to}{last reference date for the "TX_ML" determination written in
ISO8601 format (i.e. "yyyy-mm-dd"). The default is today.}

\item{region}{a character vector specifying the "State" of interest.
The default utilizes all the states in the dataframe.}

\item{site}{a character vector of at least length 1. Default is to utilize all
the facilities contained in the dataframe.}
}
\value{

}
\description{
Subset clients who became inactive (IIT) within a given period
}
\examples{
file_path <- "C:/Users/stephenbalogun/Documents/My R/tidyndr/ndr_example.csv"
ndr_example <- read_ndr(file_path)
tx_ml(ndr_example)

# Find clients who were inactive at the end of Q1 of FY21
tx_ml(ndr_example, to = "2020-12-31")
}
