% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_mmd.R
\name{tx_mmd}
\alias{tx_mmd}
\title{Subset active clients based on months of ARV Dispensed}
\usage{
tx_mmd(
  data,
  month = c(3, 4, 5, 6),
  region = unique(data$state),
  site = unique(data$facility)
)
}
\arguments{
\item{data}{an ndr dataframe imported using the \code{read_ndr()}.}

\item{month}{the number(s) of month of interest of ARV dispensed (rounded to the nearest who number).
The default is to subset active clients who had 3 - 6 months of ARV dispensed.}

\item{region}{a character vector specifying the "State" of interest.
The default utilizes all the states in the dataframe.}

\item{site}{a character vector of at least length 1. Default is to utilize all
the facilities contained in the dataframe.}
}
\value{

}
\description{
Subset active clients based on months of ARV Dispensed
}
\examples{
file_path <- "C:/Users/stephenbalogun/Documents/My R/tidyndr/ndr_example.csv"
ndr_example <- read_ndr(file_path)
tx_mmd(ndr_example)

# subset active clients who had 2 or 4 months of ARV dispensed at last encounter
tx_mmd(ndr_example, month = c(2, 4))
}
