% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_new.R
\name{tx_new}
\alias{tx_new}
\title{Subset clients starting ART within a particular period.}
\usage{
tx_new(
  data,
  from = fy_start,
  to = Sys.Date(),
  region = unique(data$state),
  site = unique(data$facility)
)
}
\arguments{
\item{data}{an ndr dataframe imported using the \code{read_ndr()}.}

\item{from}{reference date to start the determination of the "TX_NEW"
in ISO8601 format (i.e. "yyyy-mm-dd"). The default is to start at the
beginning of the current Fiscal Year (i.e. 1st of October).}

\item{to}{last reference date for the "TX_NEW" determination written in
ISO8601 format (i.e. "yyyy-mm-dd"). The default is today.}

\item{region}{a character vector specifying the "State" of interest.
The default utilizes all the states in the dataframe.}

\item{site}{a character vector of at least length 1. Default is to utilize all
the facilities contained in the dataframe.}
}
\value{

}
\description{
Subset clients starting ART within a particular period.
}
\examples{
file_path <- "C:/Users/stephenbalogun/Documents/My R/tidyndr/ndr_example.csv"
ndr_example <- read_ndr(file_path)

tx_new(ndr_example)

# generate the TX_NEW for a specific state (State 1)
tx_new(ndr_example, region = "State 1")

## Determine the TX_NEW for Quarter 1 of FY21 for State 2
ndr_example \%>\%
  tx_new(
    from = "2020-10-01",
    to = "2020-12-31",
    region = "State 2"
  )
}
