% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_pvls_den.R
\name{tx_pvls_den}
\alias{tx_pvls_den}
\title{Subset clients who have a documented viral load result within the last 1 year}
\usage{
tx_pvls_den(
  data,
  reference = Sys.Date(),
  region = unique(data$state),
  site = unique(data$facility)
)
}
\arguments{
\item{data}{an ndr dataframe imported using the `read_ndr()}

\item{reference}{date, provided in ISO8601 format ("yyyy-mm-dd"), used to
determine clients who are eligible for viral load and should have a
documented result. The default is the date of analysis (today).
Repeat viral load is calculated for adults of age 20 years annually and every
6 months for clients of 19 years and below.}

\item{region}{a character vector specifying the "State" of interest.
The default utilizes all the states in the dataframe.}

\item{site}{a character vector of at least length 1. Default is to utilize all
the facilities contained in the dataframe.}
}
\value{

}
\description{
Subset clients who have a documented viral load result within the last 1 year
}
\examples{
file_path <- "C:/Users/stephenbalogun/Documents/My R/tidyndr/ndr_example.csv"
ndr_example <- read_ndr(file_path)

tx_pvls_den(ndr_example)

# Determine clients who are virally suppressed for two state at the end of Q1
ndr_example \%>\%
  tx_pvls_den(
    reference = "2020-12-31",
    region = c("State 1", "State 2")
  )
}
