% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_rtt.R
\name{tx_rtt}
\alias{tx_rtt}
\title{Subset rows of clients who were previously inactive but now active}
\usage{
tx_rtt(
  old_data,
  new_data,
  region = unique(new_data$state),
  site = unique(new_data$facility)
)
}
\arguments{
\item{old_data}{the initial dataframe containing the list of clients who have been previously inactive.}

\item{new_data}{the current datafame where changes in current treatment status will be checked.}

\item{region}{a character vector specifying the "State" of interest.
The default utilizes all the states in the dataframe.}

\item{site}{a character vector of at least length 1. Default is to utilize all
the facilities contained in the dataframe.}
}
\value{

}
\description{
Subset rows of clients who were previously inactive but now active
}
\examples{
file_path <- "C:/Users/stephenbalogun/Documents/My R/tidyndr/ndr_example.csv"
ndr_old <- read_ndr(file_path)
ndr_new <- read_ndr(file_path)
tx_rtt(ndr_old, ndr_new)

# Determine RTT for a particular state
tx_rtt(ndr_old, ndr_new, region = "State 1")
}
