% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_vl_eligible.R
\name{tx_vl_eligible}
\alias{tx_vl_eligible}
\title{Subset clients who are eligible for viral load}
\usage{
tx_vl_eligible(
  data,
  reference = Sys.Date(),
  region = unique(data$state),
  site = unique(data$facility)
)
}
\arguments{
\item{data}{an ndr dataframe imported using the `read_ndr()}

\item{reference}{date, provided in ISO8601 format ("yyyy-mm-dd"), used to
determine clients who are eligible for viral load. The default is to subset clients
who have been on medication for at least 6 months
(ART start date is at least 6 months ago).}

\item{region}{a character vector specifying the "State" of interest.
The default utilizes all the states in the dataframe.}

\item{site}{a character vector of at least length 1. Default is to utilize all
the facilities contained in the dataframe.}
}
\value{

}
\description{
Subset clients who are eligible for viral load
}
\examples{
file_path <- "C:/Users/stephenbalogun/Documents/My R/tidyndr/ndr_example.csv"
ndr_example <- read_ndr(file_path)

tx_vl_eligible(ndr_example)

# Determine clients who are going to be eligible for VL by the end of Q2 of FY21
ndr_example \%>\%
tx_vl_eligible(reference = "2021-03-31")

}
