% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ndr-null.R
\name{ndr-null-op}
\alias{ndr-null-op}
\alias{\%||\%}
\title{Borrowed NULL}
\usage{
x \%||\% y
}
\arguments{
\item{x}{value set to a default value of null}

\item{y}{value to be used with x is null}
}
\value{
No return value, called for side effects
}
\description{
Borrowed NULL
}
\keyword{internal}
