% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ndr.R
\name{read_ndr}
\alias{read_ndr}
\title{Read NDR "csv" file}
\usage{
read_ndr(path, file_type = "current", cols = ndr_cols, ...)
}
\arguments{
\item{path}{Path to the csv file on computer.}

\item{file_type}{One of three options to specify if an old, a current version
or a new(er) version of the ndr line-list will be used. The old version was changed
to the current version at about the beginning of Fiscal Year 2021.}

\item{cols}{Sets the column types so that the columns are assigned the
appropriate class. You can supply this argument following the instructions
in \code{?vroom::cols} documentation.}

\item{...}{passes other arguments supplied to the \code{vroom::vroom()} used
behind the hood.}
}
\value{
read_ndr
}
\description{
Import your NDR patient-level line-list downloaded as ".csv" format from the
NDR front-end into R in a nicely formatted table.
}
\examples{
# Read \code{ndr_example.csv} from a path
file_path <- system.file("extdata", "ndr_example.csv", package = "tidyndr")
read_ndr(file_path)

## Not run:
# Read using a link to the NDR csv file on the internet
file_path <- "https://raw.githubusercontent.com/stephenbalogun/example_files/main/ndr_example.csv"
read_ndr(file_path)
}
