% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ndr.R
\name{read_ndr}
\alias{read_ndr}
\title{Read NDR "csv" file}
\usage{
read_ndr(path, suppress = FALSE, cols = ndr_cols, ...)
}
\arguments{
\item{path}{Path to the csv file on computer.}

\item{suppress}{Either TRUE or FALSE, The default, FALSE, instructs
\code{read_ndr()} to print warnings when using older versions of the ndr file.}

\item{cols}{Sets the column types so that the columns are assigned the
appropriate class. You can supply this argument following the instructions
in \code{?vroom::cols} documentation.}

\item{...}{passes other arguments supplied to the \code{vroom::vroom()} used
behind the hood.}
}
\value{

}
\description{
Import your NDR patient-level line-list downloaded as ".csv" format from the
NDR front-end into R in a nicely formatted table.
}
\examples{
# Read \code{ndr_example.csv} from a path
file_path <- system.file("extdata", "ndr_example.csv", package = "tidyndr")
read_ndr(file_path)

## Not run:
# Read using a link to the NDR csv file on the internet
file_path <- "https://raw.githubusercontent.com/stephenbalogun/example_files/main/ndr_example.csv"
read_ndr(file_path)
}
