% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyndr-package.R
\docType{package}
\name{tidyndr}
\alias{tidyndr}
\alias{tidyndr-package}
\title{tidyndr: A package for analyzing the de-identified patient-level data from the
Nigeria National Data Repository (NDR)}
\description{
To learn more about tidyndr, refer to the vignettes: \code{browseVignettes(package = "tidyndr")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stephenbalogun/tidyndr}
  \item Report bugs at \url{https://github.com/stephenbalogun/tidyndr/issues}
}

}
\author{
\strong{Maintainer}: Stephen Balogun \email{stephentaiyebalogun@gmail.com} (\href{https://orcid.org/0000-0002-9928-3703}{ORCID})

Other contributors:
\itemize{
  \item Scholastica Olanrewaju [contributor]
  \item Oluwaseun Okunuga [contributor]
  \item Temitope Kolade [contributor]
  \item Geraldine Abone [contributor]
}

}
\keyword{internal}
