% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_appointment.R
\name{tx_appointment}
\alias{tx_appointment}
\title{Subset rows of clients who have clinic appointment/medication pick-up within a particular period}
\usage{
tx_appointment(
  data,
  from = fy_start,
  to = Sys.Date(),
  region = unique(data$state),
  site = unique(data$facility)
)
}
\arguments{
\item{data}{an ndr dataframe imported using the `read_ndr()}

\item{from}{the start date for clients with clinic appointments in
ISO8601 format (i.e. "yyyy-mm-dd"). The default is to start at the
beginning of the current Fiscal Year (i.e. 1st of October).}

\item{to}{the end date for the appointment period of interest written
in ISO8601 format (i.e. "yyyy-mm-dd"). The default is today.}

\item{region}{a character vector specifying the "State" of interest.
The default utilizes all the states in the dataframe.}

\item{site}{a character vector of at least length 1. Default is to utilize all
the facilities contained in the dataframe.}
}
\value{

}
\description{
Subset rows of clients who have clinic appointment/medication pick-up within a particular period
}
\examples{
# Determine clients who have medication refill in Q2 of Fy21
file_path <- "C:/Users/stephenbalogun/Documents/My R/tidyndr/ndr_example.csv"
ndr_example <- read_ndr(file_path)
tx_appointment(ndr_example,
  from = "2021-01-01",
  to = "2021-03-30"
)

# Determine clients with medication refill in January 2021 for a particular facility
tx_appointment(ndr_example,
  from = "2021-01-01",
  to = "2021-01-31",
  region = "State 1",
  site = "Facility 1"
)
}
