% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_curr.R
\name{tx_curr}
\alias{tx_curr}
\title{Subset clients who are currently on treatment}
\usage{
tx_curr(data, states = regions, facilities = sites)
}
\arguments{
\item{data}{An ndr dataframe imported using the `read_ndr().}

\item{states}{The name(s) of the State(s) of interest. The default utilizes all
the states in the dataframe. If specifying more than one state, combine the
states using the \code{c()} e.g. \code{c("State 1", "State 2")}.}

\item{facilities}{The name(s) of the facilit(ies) of interest. Default is to utilize
all the facilities contained in the dataframe. If specifying more than one
facility, combine the facilities using the \code{c()} e.g.
\code{c("Facility 1", "Facility 2")}.}
}
\value{

}
\description{
\code{tx_curr} pulls up the line-list of clients who are active
on treatment using the \code{current_status_28_days} column. You can specify
the state(s) and/or facilit(ies) of interest using the \code{region} or
\code{site} arguments.
}
\examples{
tx_curr(ndr_example)

# generate the TX_CURR for two states (e.g. "State 1" and "State 2" in the ndr_example file)
tx_curr(ndr_example,
  states = c("State 1", "State 2")
)

# determine the active clients in two facilities ("Facility 1", and "Facility 2) in "State 1"
tx_curr(ndr_example,
  states = "State 1",
  facilities = c("Facility 1", "Facility 2")
)
}
