% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_ml.R
\name{tx_ml}
\alias{tx_ml}
\title{Subset clients who became inactive (IIT) within a given period}
\usage{
tx_ml(
  data,
  from = get("fy_start")(),
  to = get("Sys.Date")(),
  states = .s,
  facilities = .f
)
}
\arguments{
\item{data}{An ndr dataframe imported using the `read_ndr().}

\item{from}{The start date in ISO8601 format (i.e. "yyyy-mm-dd").
The default is to start at the beginning of the current Fiscal Year (i.e. 1st of October).}

\item{to}{The end date written in ISO8601 format (i.e. "yyyy-mm-dd").
The default is today (the date of analysis).}

\item{states}{The name(s) of the State(s) of interest. The default utilizes all
the states in the dataframe. If specifying more than one state, combine the
states using the \code{c()} e.g. \code{c("State 1", "State 2")}.}

\item{facilities}{The name(s) of the facilit(ies) of interest. Default is to utilize
all the facilities contained in the dataframe. If specifying more than one
facility, combine the facilities using the \code{c()} e.g.
\code{c("Facility 1", "Facility 2")}.}
}
\value{
tx_ml
}
\description{
\code{tx_ml} Generates clients who have become inactive over a specified
period of time. The default is to generate all clients who became inactive
in the current Fiscal Year. You can specify the period of interest
(using the \code{from} and \code{to} arguments). Used together
with \code{tx_ml_outcomes()}, generates inactive clients with a particular
outcome of interest.
}
\examples{
tx_ml(ndr_example)

# Find clients who were inactive at the end of Q1 of FY21
tx_ml(ndr_example, to = "2020-12-31")
}
