% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_ml_outcomes.R
\name{tx_ml_outcomes}
\alias{tx_ml_outcomes}
\title{Subset rows of Inactive Clients with Specific Outcome}
\usage{
tx_ml_outcomes(data, outcome)
}
\arguments{
\item{data}{An ndr dataframe imported using the `read_ndr()}

\item{outcome}{The particular outcome of interest based on options available
on the NDR ("transfer out" or "dead").}
}
\value{
tx_ml_outcomes
}
\description{
\code{tx_ml_outcomes} generates the line-list of clients based on the
outcome of interest ("dead" or "transfer out"). It should be used after
\code{tx_ml()}.
}
\examples{
tx_ml_outcomes(tx_ml(new_data = ndr_example),
  outcome = "dead"
)
}
