% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_mmd.R
\name{tx_mmd}
\alias{tx_mmd}
\title{Subset active clients based on months of ARV Dispensed}
\usage{
tx_mmd(data, month = months_dispensed, state = region, facility = site)
}
\arguments{
\item{data}{An ndr dataframe imported using the \code{read_ndr()}.}

\item{month}{The number(s) of month of interest of ARV dispensed
(rounded to the nearest who number). The default is to subset active
clients who had 3 - 6 months of ARV dispensed.}

\item{state}{The name(s) of the State(s) of interest. The default utilizes all
the states in the dataframe. If specifying more than one state, combine the
states using the \code{c()} e.g. \code{c("State 1", "State 2")}.}

\item{facility}{The name(s) of the facilit(ies) of interest. Default is to utilize
all the facilities contained in the dataframe. If specifying more than one
facility, combine the facilities using the \code{c()} e.g.
\code{c("Facility 1", "Facility 2")}.}
}
\value{

}
\description{
Generates list of clients who had 3 - 6 months of ARV dispensed during the
medication refill. You can specify the number of month(s) of ARV dispensed
by changing the \code{month} argument.
}
\examples{
tx_mmd(ndr_example)

# subset active clients who had 2 or 4 months of ARV dispensed at last encounter
tx_mmd(ndr_example,
  month = c(2, 4)
)
}
