% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_regimen.R
\name{tx_regimen}
\alias{tx_regimen}
\title{Subset Clients Based on their Current ART Regimen}
\usage{
tx_regimen(data, age_band = age_range, states = regions, facilities = sites)
}
\arguments{
\item{data}{An ndr dataframe imported using the `read_ndr()}

\item{age_band}{a numeric vector of length 2 (lower age-range and upper
age-range).}

\item{states}{The name(s) of the State(s) of interest. The default utilizes all
the states in the dataframe. If specifying more than one state, combine the
states using the \code{c()} e.g. \code{c("State 1", "State 2")}.}

\item{facilities}{The name(s) of the facilit(ies) of interest. Default is to utilize
all the facilities contained in the dataframe. If specifying more than one
facility, combine the facilities using the \code{c()} e.g.
\code{c("Facility 1", "Facility 2")}.}
}
\value{
tx_regimen
}
\description{
Generates the line-list of clients on first-line regimen who are on the choice
combination regimen for their age or weight.
}
\examples{
tx_regimen(ndr_example)
}
